Remove the need to enter api connection.
